+++ NEWS.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## Version 0.7.4 (2024-11-29)
+
+* Included changes for `CO2Storage` to not require the definition of variables that are not used.
+* Included check changes from `EnergyModelsBase`.
+* Fixed minor errors in the documentation.
+
 ## Version 0.7.3 (2024-10-16)
 
 * Minor changes to the documentation and docstrings.